the fort commit

+++ networksecurity/utils/main_utils/utils.py
@@ -1,28 +1,51 @@
 import yaml
 from networksecurity.exception.exception import NetworkSecurityException
 from networksecurity.logging.logger import logging
-import os,sys
+import os, sys
 import numpy as np
 import dill
 import pickle
 
-
 def read_yaml_file(file_path: str) -> dict:
     try:
-        with open(file_path,"rb")as yaml_file:
+        logging.info(f"Reading YAML file from {file_path}")
+        with open(file_path, "rb") as yaml_file:
             return yaml.safe_load(yaml_file)
     except Exception as e:
-        raise NetworkSecurityException(e,sys) from e
+        raise NetworkSecurityException(e, sys) from e
 
-def write_yaml_file(file_path: str,content: object, replace: bool=False)->None:
+def write_yaml_file(file_path: str, content: object, replace: bool = False) -> None:
     try:
-        if replace:
-            if os.path.exists(file_path):
-                os.remove(file_path)
-            os.makedirs(os.path.dirname(file_path),exist_ok=True)
-            with open(file_path,"w") as file:
-                yaml.dump(content,file)
-    except Exception as e:
-        raise NetworkSecurityException(e,sys)           
-    
+        if replace and os.path.exists(file_path):
+            os.remove(file_path)
 
+        os.makedirs(os.path.dirname(file_path), exist_ok=True)
+        with open(file_path, "w") as file:
+            yaml.dump(content, file)
+        logging.info(f"Wrote YAML file to {file_path}")
+    except Exception as e:
+        raise NetworkSecurityException(e, sys)
+
+def save_numpy_array_data(file_path: str, array: np.array):
+    """
+    Save numpy array data to file.
+    """
+    try:
+        logging.info(f"Saving NumPy array to {file_path}")
+        dir_path = os.path.dirname(file_path)
+        os.makedirs(dir_path, exist_ok=True)
+        with open(file_path, "wb") as file_obj:
+            np.save(file_obj, array)
+    except Exception as e:
+        raise NetworkSecurityException(e, sys) from e
+
+def save_object(file_path: str, obj: object) -> None:
+    try:
+        logging.info("Entered the save_object method of MainUtils class")
+        os.makedirs(os.path.dirname(file_path), exist_ok=True)
+        with open(file_path, "wb") as file_obj:
+            pickle.dump(obj, file_obj)
+        logging.info(f"Saved object to {file_path}")
+        logging.info("Exited the save_object method in MainUtils class")
+    except Exception as e:
+        raise NetworkSecurityException(e, sys) from e
